+++ src/pull/mod.rs
@@ -29,10 +29,10 @@ fn add_default_request_headers(req: RequestBuilder) -> RequestBuilder {
     req.header(USER_AGENT, format!("radiobrowser-api-rust/{}",pkg_version))
 }
 
-fn pull_worker(client: &Client, connection_string: String, mirrors: &Vec<String>) -> Result<(),Box<dyn Error>> {
+fn pull_worker(client: &Client, connection_string: String, mirrors: &Vec<String>, chunk_size_changes: usize) -> Result<(),Box<dyn Error>> {
     let pool = connect(connection_string)?;
     for server in mirrors.iter() {
-        let result = pull_server(client, &pool, &server);
+        let result = pull_server(client, &pool, &server, chunk_size_changes);
         match result {
             Ok(_) => {
             },
@@ -44,12 +44,12 @@ fn pull_worker(client: &Client, connection_string: String, mirrors: &Vec<String>
     Ok(())
 }
 
-pub fn start(connection_string: String, mirrors: Vec<String>, pull_interval: u64) {
+pub fn start(connection_string: String, mirrors: Vec<String>, pull_interval: u64, chunk_size_changes: usize) {
     if mirrors.len() > 0 {
         thread::spawn(move || {
             let client = Client::new();
             loop {
-                let result = pull_worker(&client, connection_string.clone(), &mirrors);
+                let result = pull_worker(&client, connection_string.clone(), &mirrors, chunk_size_changes);
                 match result {
                     Ok(_) => {
                     },
@@ -94,11 +94,11 @@ fn pull_history(client: &Client, server: &str, api_version: u32, lastid: Option<
     }
 }
 
-fn pull_checks(client: &Client, server: &str, api_version: u32, lastid: Option<String>) -> Result<Vec<StationCheck>, Box<dyn std::error::Error>> {
+fn pull_checks(client: &Client, server: &str, api_version: u32, lastid: Option<String>, chunk_size_changes: usize) -> Result<Vec<StationCheck>, Box<dyn std::error::Error>> {
     trace!("Pull checks from '{}' (API: {}) ..", server, api_version);
     let path = match lastid {
-        Some(id) => format!("{}/json/checks?lastcheckuuid={}",server, id),
-        None => format!("{}/json/checks",server),
+        Some(id) => format!("{}/json/checks?lastcheckuuid={}&limit={}",server, id, chunk_size_changes),
+        None => format!("{}/json/checks?limit={}",server,chunk_size_changes),
     };
     trace!("{}", path);
     let result = add_default_request_headers(client.get(&path)).send()?;
@@ -184,7 +184,7 @@ fn pull_stations(client: &Client, server: &str, api_version: u32) -> Result<Vec<
     }
 }
 
-fn pull_server(client: &Client, connection_new: &Box<dyn DbConnection>, server: &str) -> Result<(),Box<dyn std::error::Error>> {
+fn pull_server(client: &Client, connection_new: &Box<dyn DbConnection>, server: &str, chunk_size_changes: usize) -> Result<(),Box<dyn std::error::Error>> {
     let insert_chunksize = 2000;
     let mut station_change_count = 0;
     let mut station_check_count = 0;
@@ -211,9 +211,9 @@ fn pull_server(client: &Client, connection_new: &Box<dyn DbConnection>, server: 
         }
     }
 
-    {
+    loop {
         let lastcheckid = connection_new.get_pull_server_lastcheckid(server)?;
-        let mut list_checks = pull_checks(client, server, api_version, lastcheckid)?;
+        let mut list_checks = pull_checks(client, server, api_version, lastcheckid, chunk_size_changes)?;
         let len = list_checks.len();
 
         trace!("Incremental checks sync ({})..", len);
@@ -279,6 +279,9 @@ fn pull_server(client: &Client, connection_new: &Box<dyn DbConnection>, server: 
                 let (_ignored_uuids_check_existing, _ignored_uuids_station_missing, inserted) = connection_new.insert_checks(checks_ignored_station_missing)?;
                 trace!("Inserted checks ({})..", inserted.len());
             }
+        }
+        if len < chunk_size_changes {
+            break;
         }
     }